requirements.txt
pandas
numpy
matplotlib
plotly
json5
python-dotenv
pyyaml
google-generativeai
requests
openai
langchain
rich
loguru
click
pytest